[publishing] swipe 훅 분리 및 스타일 변경
